chore: change value attribute type to float

+++ src/main/java/dlt/client/tangle/hornet/model/transactions/reputation/Evaluation.java
@@ -7,24 +7,24 @@ import dlt.client.tangle.hornet.model.transactions.TargetedTransaction;
  * Transaction evaluating the service received.
  *
  * @author Allan Capistrano
- * @version 1.0.0
+ * @version 1.0.1
  */
 public class Evaluation extends TargetedTransaction {
 
-  private final int value;
+  private final float value;
 
   public Evaluation(
     String source,
     String target,
     String group,
     TransactionType type,
-    int value
+    float value
   ) {
     super(source, group, type, target);
     this.value = value;
   }
 
-  public int getValue() {
+  public float getValue() {
     return value;
   }
 }
